Expand NKIGym optimization coverage with frozen NAKB benchmarks and replayable kernel ladders

+++ nkigym/src/nkigym/ops/grouped_range_select_reduce.py
@@ -40,7 +40,7 @@ class NKIGroupedRangeSelectReduce(NKIOp):
     CODEGEN_ONLY_KWARGS: ClassVar[frozenset[str]] = frozenset({"groups", "queries", "tiles", "partitions", "width"})
     SPLIT_OFFSET_KWARGS: ClassVar[dict[str, tuple[str, str]]] = {"F": ("range_start", "on_true_tile")}
     OUTPUT_DTYPES: ClassVar[dict[str, str]] = {"dst": "float32", "reduce_res": "float32"}
-    OUTPUT_STORAGE_DTYPES: ClassVar[dict[str, str]] = {"dst": "float32", "reduce_res": "float32"}
+    OUTPUT_STORAGE_DTYPES: ClassVar[dict[str, str]] = {"dst": "bfloat16", "reduce_res": "float32"}
     OUTPUT_LOCATION: ClassVar[str] = "sbuf"
 
     def __init__(self, groups: int, queries: int, tiles: int, partitions: int, width: int) -> None:
